Implement current_user into redux inside MyChart.js
