+++ src/actions/types.js
@@ -1,5 +1,3 @@
 export const FETCH_RESULTS = 'FETCH_RESULTS';
 export const FETCH_DETAIL_RESULT = 'FETCH_DETAIL_RESULT';
 export const FETCH_CURRENT_USER = 'FETCH_CURRENT_USER';
-export const POST_NEW_USER = 'POST_NEW_USER';
-export const POST_NEW_RESULTS = 'POST_NEW_RESULTS';
